+++ bk2ch08p439searchableTable2/ch21p718sections/RootViewController.m
@@ -84,7 +84,7 @@
 
 - (NSInteger)tableView:(UITableView *)tableView
  numberOfRowsInSection:(NSInteger)section {
-    NSLog(@"number of rows in section %d", section);
+    NSLog(@"number of rows in section %ld", (long)section);
     NSArray* model =
     (tableView == self.sdc.searchResultsTableView) ?
     self.filteredSectionData : self.sectionData;
@@ -93,7 +93,7 @@
 
 - (UITableViewCell *)tableView:(UITableView *)tableView
          cellForRowAtIndexPath:(NSIndexPath *)indexPath {
-    NSLog(@"cell for section %d row %d", indexPath.section, indexPath.row);
+    NSLog(@"cell for section %ld row %ld", (long)indexPath.section, (long)indexPath.row);
     UITableViewCell *cell =
     [self.tableView dequeueReusableCellWithIdentifier:@"Cell"
                                     forIndexPath:indexPath];
@@ -115,7 +115,7 @@
 }
 
 - (UIView *)tableView:(UITableView *)tableView viewForHeaderInSection:(NSInteger)section {
-    NSLog(@"view for header %d", section);
+    NSLog(@"view for header %ld", (long)section);
     
     UITableViewHeaderFooterView* h =
     [tableView dequeueReusableHeaderFooterViewWithIdentifier:@"Header"];
